Observer and TableView to Record Data
Added observer pattern for particles leaving the screen.
This adds a record to a data collection class.  The idea is to have this
class add data to TableView when a particle leaves the screen.  While
this feature is not functional currently, I am essentially backing up a
working program in the event that my refactoring into a MVC needs to be
reverted to a previous state.